File structure adjusted to be more organised
