+++ Public/Script/CheckLetterInWord.js
