+++ Public/View/SinglePlayer.html
@@ -6,7 +6,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <title>Guessle</title>
 
-        <link rel="stylesheet" href="style.css">
+        <link rel="stylesheet" href="../Style/style.css">
     </head>
     <body>
         <h1> Guessle Game </h1>
@@ -18,8 +18,8 @@
                 src="https://code.jquery.com/jquery-3.6.0.min.js"
             integrity="sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4="
             crossorigin="anonymous"></script>
-        <script type="module" src="boardScript.js" ></script>
-        <script type="module" src="inputLetters.js"></script>
+        <script type="module" src="../Script/boardScript.js" ></script>
+        <script type="module" src="../Script/inputLetters.js"></script>
     <div id="currentBlock"></div>
     </body>
 </html>
