--- words.js
@@ -1,5 +1,0 @@
-export const WORDS = [
-    'Secnd',
-    'Third',
-    'Forth',
-    'Phith']
